requirements.txt
opencv-python
gunicorn
Flask
Flask-Cors
Jinja2
Werkzeug
distribute
Pillow
Scrapy
Twisted
